Add registerController to userController
