+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Support\Facades\Auth;
+use App\Providers\RouteServiceProvider;
 
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -15,13 +16,17 @@ class UserController extends Controller
             $user = Auth::user();
         }
 
-
         return view('users.profil', compact('user'));
     }
 
     public function login()
     {
         return view('users.login');
+    }
+
+    public function register()
+    {
+        return view('users.register');
     }
 
     public function logout()
@@ -45,5 +50,34 @@ class UserController extends Controller
         } else {
             return redirect()->route('login')->with('error', 'Username or password is incorrect.');
         }
+    }
+
+    public function createUser(Request $request)
+    {
+        $request->validate([
+            'firstname' => 'required|string|max:255',
+            'lastname' => 'required|string|max:255',
+            'email' => 'required|email|unique:users,email',
+            'username' => 'required|string|max:255|unique:users,username',
+            'password' => 'required|string|min:8',
+        ]);
+
+        // Create new user
+        $user = new User();
+        $user->firstname = $request->input('firstname');
+        $user->lastname = $request->input('lastname');
+        $user->email = $request->input('email');
+        $user->username = $request->input('username');
+        $user->password = bcrypt($request->input('password'));
+
+
+        $user->name = $request->input('firstname');
+
+
+        $user->save();
+
+        auth()->login($user);
+
+        return redirect()->route('home')->with('success', 'Account created successfully!');
     }
 }
