+++ resources/views/layout/app.blade.php
@@ -19,10 +19,10 @@
             </button>
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav">
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ route('lists.index') }}">Lists</a>
-                    </li>
                     @auth
+                        <li class="nav-item">
+                            <a class="nav-link" href="{{ route('lists.index') }}">Lists</a>
+                        </li>
                         <li class="nav-item">
                             <a class = "nav-link" href="{{ url('/profil') }}" class="font-semibold text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-white focus:outline focus:outline-2 focus:rounded-sm focus:outline-red-500">Profil</a>
                         </li>
